+++ CHANGELOG.rst
@@ -9,6 +9,10 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+Changed
+^^^^^^^
+* Change toml parser to tomli
+
 Removed
 ^^^^^^^
 * Removed changelog from the published wheel